refactor: restructuring the folder names
