+++ src/infraestructure/factories/usecases/activity/make-db-add-activity.ts
@@ -2,8 +2,8 @@ import {
   CreateActivity,
   CreateActivityUseCase,
 } from "../../../../application/usecases/activity/create-activity-use-case";
-import { ActivityRepositoryImpl } from "../../../db/repository/activity/ActivitRepositoryImpl";
-import { logging } from "../../../logger/logging";
+import { ActivityRepositoryImpl } from "../../../database/repositories/activity/activity-repository";
+import { logging } from "../../../logging/logging";
 
 export const makeDbAddActvity = (): CreateActivity => {
   const activiyRepository = new ActivityRepositoryImpl();
