+++ src/infraestructure/factories/usecases/activity/make-db-delete-activity.ts
@@ -2,8 +2,8 @@ import {
   DeleteActivity,
   DeleteActivityUseCase,
 } from "../../../../application/usecases/activity/delete-activity-use-case";
-import { ActivityRepositoryImpl } from "../../../db/repository/activity/ActivitRepositoryImpl";
-import { logging } from "../../../logger/logging";
+import { ActivityRepositoryImpl } from "../../../database/repositories/activity/activity-repository";
+import { logging } from "../../../logging/logging";
 
 export const makeDbDeleteActivity = (): DeleteActivity => {
   const activityRepository = new ActivityRepositoryImpl();
