+++ src/infraestructure/factories/usecases/activity/make-db-update-activity.ts
@@ -2,8 +2,8 @@ import {
   UpdateActivity,
   UpdateActivityUseCase,
 } from "../../../../application/usecases/activity/update-activity-use-case";
-import { ActivityRepositoryImpl } from "../../../db/repository/activity/ActivitRepositoryImpl";
-import { logging } from "../../../logger/logging";
+import { ActivityRepositoryImpl } from "../../../database/repositories/activity/activity-repository";
+import { logging } from "../../../logging/logging";
 
 export const makeDbUpdateActivity = (): UpdateActivity => {
   const activityRepository = new ActivityRepositoryImpl();
