+++ src/presentation/routes/routes.ts
@@ -1,7 +1,8 @@
 import { Router } from "express";
-import { userRoutes } from "./users/user-routes";
+import { userManagementRoutes } from "./users/user-routes";
 import { authRoutes } from "./authentication/auth-routes";
 
 export const router = Router();
-router.use("/users", userRoutes);
+
+router.use("/users", userManagementRoutes);
 router.use("/authentication/user", authRoutes);
