+++ src/presentation/routes/users/user-routes.ts
@@ -2,6 +2,6 @@ import { Router } from "express";
 import { adaptRoute } from "../../../adapters/router-adapter";
 import { makeAddUserController } from "../../../infraestructure/factories/controllers/user/make-add-controller";
 
-export const userRoutes = Router();
+export const userManagementRoutes = Router();
 
-userRoutes.post("/", adaptRoute(makeAddUserController()));
+userManagementRoutes.post("/", adaptRoute(makeAddUserController()));
